# src/main/resources/swagger/example/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8080/api"
    }
  ],
  "info": {
    "description": "This is a simple example API for Helpdesk",
    "version": "1.0.0",
    "title": "Simple Helpdesk API",
    "contact": {
      "email": "mentor@novahub.vn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/admin/skills": {
      "post": {
        "tags": [
          "Skill"
        ],
        "summary": "Create new skill (ADMIN)",
        "description": "Create new skill (ADMIN)",
        "parameters": [
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Java"
                  },
                  "category_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get all skills successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill",
                  "properties": {
                    "category": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/skills"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/skills"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532462094.006"
                    },
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Category not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "CategoryNotFoundException with id = 1"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/skills"
                    }
                  }
                }
              }
            }
          },
          "406": {
            "description": "Not acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 406
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "enum": [
                            "name is not empty"
                          ],
                          "example": "name is not empty"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "SkillValidationException with number of errors = 1"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/skills"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Skill is exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 409
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Skill is exist"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "SkillIsExistException with name = Java"
                    },
                    "path": {
                      "type": "string",
                      "example": "/admin/skills"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/skills/{id}": {
      "get": {
        "tags": [
          "Skill"
        ],
        "summary": "Get a skill by id",
        "description": "Get a skill by id",
        "parameters": [
          {
            "in": "header",
            "name": "access_token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "dff4d0d4891bfd8bafdb859dac4cbccb218bcf5adc9887a627c6c159f76ef0b9"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Skill",
                  "properties": {
                    "category": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token is expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": 1532589762.947
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Token is expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "TokenIsExpiredException with token = 579989c508208d49c978e6cb99383688e6f1fedabfc22c27d3542ab0341d3b13"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 401
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid token"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access is denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532506012.933"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 403
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access is denied"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Access is denied"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "skill not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "example": "1532462094.006"
                    },
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Skill not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "SkillNotFoundException with id = {id}"
                    },
                    "path": {
                      "type": "string",
                      "example": "/skills/{id}"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Skill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Java"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 18:39:48"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 18:39:48"
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Back-end"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2018-07-24 17:50:45"
          }
        }
      },
...
    }
  }
}